+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - fixed `addValidator()` method
   - added support for `requirementsTransformer` for custom validators
   - updated jQuery needed version from README and doc
+  - fixed case when a multiple item were dynamically removed from DOM (#634)
 
 ## 2.0.0